policy

+++ app/Http/Controllers/admin/RolesController.php
@@ -12,7 +12,7 @@ class RolesController extends Controller
 {
     public function index()
     {
-        // Gate::authorize('roles.view-any');
+        Gate::authorize('roles.view-any');
 
         $roles = Role::paginate();
         return view('admin.roles.index', compact('roles'));
@@ -25,7 +25,7 @@ class RolesController extends Controller
      */
     public function create()
     {
-        // Gate::authorize('roles.create');
+        Gate::authorize('roles.create');
         return view('admin.roles.create', [
             'role' => new Role(),
         ]);
@@ -39,7 +39,7 @@ class RolesController extends Controller
      */
     public function store(Request $request)
     {
-        // Gate::authorize('roles.create');
+        Gate::authorize('roles.create');
         $request->validate([
             'name' => 'required',
             'abilities' => 'required|array',
@@ -72,7 +72,7 @@ class RolesController extends Controller
      */
     public function edit($id)
     {
-        // Gate::authorize('roles.update');
+        Gate::authorize('roles.update');
         $role = Role::findOrFail($id);
 
         return view('admin.roles.edit', compact('role'));
@@ -87,7 +87,7 @@ class RolesController extends Controller
      */
     public function update(Request $request, $id)
     {
-        // Gate::authorize('roles.update');
+        Gate::authorize('roles.update');
         $request->validate([
             'name' => 'required',
             'abilities' => 'required|array',
@@ -107,7 +107,7 @@ class RolesController extends Controller
      */
     public function destroy($id)
     {
-        // Gate::authorize('roles.delete');
+        Gate::authorize('roles.delete');
 
         $role = Role::findOrFail($id);
         $role->delete();
